refactor CryptorRuleHandle
